+++ CHANGES.md
@@ -6,6 +6,14 @@ Version 0.1.1
 
 To be released.
 
+### @upyo/smtp
+
+ -  Fixed CJK character encoding corruption in SMTP transport HTML emails.
+    Korean, Japanese, and Chinese characters are now properly encoded using
+    UTF-8 quoted-printable encoding.  [[#4]]
+
+[#4]: https://github.com/dahlia/upyo/issues/4
+
 
 Version 0.1.0
 -------------